requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.0
opencv-python==4.8.1.78
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.7.1
yarl==1.9.2
